requirements.txt
praw
python-dotenv
tiktoken
openai